introduce encryption package
