+++ apps/extension/package.json
@@ -35,6 +35,7 @@
     "@penumbra-zone/wasm": "53.0.0",
     "@radix-ui/react-icons": "^1.3.2",
     "@repo/context": "workspace:*",
+    "@repo/encryption": "workspace:*",
     "@repo/storage-chrome": "workspace:*",
     "@repo/ui": "workspace:*",
     "@repo/wallet": "workspace:*",
