+++ packages/encryption/src/key-stretching.ts
@@ -1,0 +1,25 @@
+// Hash a password with PBKDF2 using a provided salt.
+// Meant to hinder brute force or dictionary attacks.
+export const keyStretchingHash = async (password: string, salt: Uint8Array): Promise<CryptoKey> => {
+  const enc = new TextEncoder();
+  const importedKey = await crypto.subtle.importKey(
+    'raw',
+    enc.encode(password),
+    { name: 'PBKDF2' },
+    false,
+    ['deriveBits', 'deriveKey'],
+  );
+
+  return crypto.subtle.deriveKey(
+    {
+      name: 'PBKDF2',
+      salt,
+      iterations: 210000,
+      hash: 'SHA-512',
+    },
+    importedKey,
+    { name: 'AES-GCM', length: 256 },
+    true,
+    ['encrypt', 'decrypt'],
+  );
+};
