+++ packages/storage-chrome/package.json
@@ -30,6 +30,7 @@
     "@penumbra-zone/getters": "28.0.0",
     "@penumbra-zone/types": "36.0.0",
     "@penumbra-zone/wasm": "53.0.0",
+    "@repo/encryption": "workspace:*",
     "@types/chrome": "0.0.270",
     "@types/lodash": "^4.17.13"
   }
