--- packages/wallet/src/encryption/index.ts
@@ -1,2 +1,0 @@
-export { Key, KeyPrint } from './key';
-export { Box } from './box';
